% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyze_alleles}
\alias{analyze_alleles}
\title{Pseudobulk analysis of allele frequencies}
\usage{
analyze_alleles(crispresso_allele_df, ignore_subs = TRUE)
}
\arguments{
\item{crispresso_allele_df}{Cleaned allele dataframe generated by `clean_crispresso_output()`}

\item{ignore_subs}{Ignore substitutions since CRISPR-Cas9 only produces indels. Set to FALSE if running on base/prime editing data. Default is TRUE.}
}
\value{
A dataframe with alleles and their total read counts
}
\description{
Perform pseudobulk analysis to identify most common alleles or indel patterns in data while ignoring cell barcodes. Only used for CRISPR-Cas9 for now.
}
\examples{
data(cocktail)
mydata <- clean_crispresso_output(cocktail)
mytab <- analyze_alleles(mydata)

}
