% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_crispresso_output}
\alias{clean_crispresso_output}
\title{Clean CRISPResso output}
\usage{
clean_crispresso_output(
  crispresso_allele_df,
  allele_read_coverage_threshold = 2,
  ignore_subs = TRUE,
  af_cutoff = 1,
  target_id = NULL
)
}
\arguments{
\item{crispresso_allele_df}{Dataframe containing allele read counts for each cell barcode. Generated by running CRISPResso on each cell barcode fastq file.}

\item{allele_read_coverage_threshold}{Integer value indicating the limit of detection for rare indels. Alleles/indels with read counts less than this value are assumed to be sequencing errors and will be removed from data. Default is 2.}

\item{ignore_subs}{Ignore substitutions since CRISPR-Cas9 only produces indels. Set to FALSE if running on base/prime editing data. Default is TRUE.}

\item{af_cutoff}{Integer corresponding to minimum pseudobulk percent that an allele must surpass in order to be analyzed downstream. Default is 1.}

\item{target_id}{Name of loci being targeted; typically name/ID of gRNA used when running CRISPResso.}
}
\value{
A dataframe with cell barcodes, allele read counts, and allele sequences used for downstream genotyping analysis.
}
\description{
Prepare cell allele frequency dataframe for genotyping analysis
}
\examples{
data(cocktail)
mydata <- clean_crispresso_output(cocktail)

}
