% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_homs}
\alias{find_homs}
\title{Identify homozygous cells}
\usage{
find_homs(
  cleaned_crispresso_allele_df,
  output_model_fit = FALSE,
  hard_thr = 75,
  seed = 92365,
  ...
)
}
\arguments{
\item{cleaned_crispresso_allele_df}{Cleaned allele dataframe generated by `clean_crispresso_output()`}

\item{output_model_fit}{Logical indicating if the model fit should be included in function output. Default is FALSE.}

\item{hard_thr}{If largest cluster mean is less than this percent value a warning is thrown suggesting absence of homozygous population in sample. Defaults to 75.}

\item{seed}{Seed for GMM initialization}

\item{...}{Additional arguments passed to `mclust::Mclust()`}
}
\value{
A list with cell genotype counts, barcodes of homozygous cells, input dataframe, and optional GMM model fit object
}
\description{
Find homozygous cells from CRISPResso output by fitting Gaussian mixture model to dominant allele frequency distribution. Homozygous cells include unedited WT and biallelic edited cells.
}
\examples{

data(cocktail)
mydata <- clean_crispresso_output(cocktail)
myhoms <- find_homs(mydata)

}
