% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_putative_hets}
\alias{find_putative_hets}
\title{Separate putative heterozygous cells from transparent multiplets}
\usage{
find_putative_hets(
  cleaned_crispresso_allele_df,
  hom_object = NULL,
  output_hom = FALSE,
  output_model_fit = FALSE,
  seed = 3594,
  ...
)
}
\arguments{
\item{cleaned_crispresso_allele_df}{Cleaned allele dataframe generated by `clean_crispresso_output()`}

\item{hom_object}{Optional. List object output by *find_homs()*. If no object is provided, `find_homs()` will automatically be executed.}

\item{output_hom}{Logical indicating if `find_homs()` model fit should be output.}

\item{output_model_fit}{Logical indicating if the heterozygous model fit should be included in function output. Default is FALSE.}

\item{seed}{Seed for GMM initialization}

\item{...}{Additional arguments passed to `mclust::Mclust()`}
}
\value{
A list with cell genotype counts, barcodes of putative heterozygous cells, input dataframe, and optional GMM model fit object
}
\description{
Find putative heterozygous cells from CRISPResso output by fitting Gaussian mixture model to the log2 ratio of the 2nd and 3rd (noise) allele frequencies. Cells with no noise allele are whitelisted as true heterozygous or homozygous if they were already identified in *find_homs()*. Cells with higher log2 ratios have lower noise allele counts and are more likely to be true heterozygous cells instead of multiplets.
}
\examples{
data(cocktail)
mydata <- clean_crispresso_output(cocktail)
myhoms <- find_homs(mydata)
myput_hets <- find_putative_hets(cleaned_crispresso_allele_df, myhoms)

}
