% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{refine_hets}
\alias{refine_hets}
\title{Separate confident heterozygous cells from opaque multiplets}
\usage{
refine_hets(
  putative_het_object,
  cleaned_crispresso_allele_df,
  dimred = TRUE,
  npcs = 2,
  output_model_fit = FALSE,
  seed = 26433,
  ...
)
}
\arguments{
\item{putative_het_object}{List object output by `find_putative_hets()`}

\item{cleaned_crispresso_allele_df}{Cleaned allele dataframe generated by `clean_crispresso_output()`}

\item{dimred}{Logical indicating if PCA should be run on log-transformed allele read counts prior to fitting GMM. Defaults to TRUE.}

\item{npcs}{Integer value indicating how many principal components should be passed to GMM. Cannot be more than number of unique allele features.}

\item{output_model_fit}{Logical indicating if the heterozygous model fit should be included in function output. Default is FALSE.}

\item{seed}{Seed for GMM initialization}

\item{...}{Additional arguments passed to `mclust::Mclust()`}
}
\value{
A list with cell genotype counts, barcodes of putative heterozygous cells, input dataframe, and optional GMM model fit object
}
\description{
Identify opaque multiplets which have allele frequency distributions similar to true heterozygous cells (no noise alleles). These droplets tend to have indel profile combinations that are rare based on their correlation structure across entire population of cells. A multivariate Gaussian mixture model is fit to all allele counts to separate true heterozygous cells from opaque multiplets.
}
\examples{
data(cocktail)
mydata <- clean_crispresso_output(cocktail)
myhoms <- find_homs(mydata)
myput_hets <- find_putative_hets(cleaned_crispresso_allele_df, myhoms)
myhets <- refine_hets(myput_hets, cleaned_crispresso_allele_df)

}
